refactor: 🔨 file preview component
