+++ packages/auto-design-system/src/components/common/FilePreview/FilePreview.tsx
@@ -1,176 +1,27 @@
 import { OffchainMetadata } from '@autonomys/auto-dag-data'
 import { NetworkId } from '@autonomys/auto-utils'
-import {
-  ArrowTopRightOnSquareIcon,
-  CodeBracketIcon,
-  DocumentIcon,
-  DocumentTextIcon,
-  LockClosedIcon,
-  MusicalNoteIcon,
-} from '@heroicons/react/24/outline'
-import React, { useCallback, useMemo, useState } from 'react'
-import { cn } from '../../../utils/cn'
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from '../../astral/Modal/Modal'
+import React, { useMemo, useState } from 'react'
 import { Arguments } from './Arguments'
+import { AudioPlayer } from './AudioPlayer'
+import { DecryptionErrorDisplay } from './DecryptionErrorDisplay'
+import { DirectGatewayLink } from './DirectGatewayLink'
+import { EncryptedFilePrompt } from './EncryptedFilePrompt'
+import { ErrorDisplay } from './ErrorDisplay'
 import { FolderPreview } from './FolderPreview'
+import { ImageViewer } from './ImageViewer'
+import { LoadingSpinner } from './LoadingSpinner'
 import { NoPreviewAvailable } from './NoPreviewAvailable'
-
-const ImageViewer = ({ src, alt }: { src: string; alt?: string }) => {
-  return (
-    <div className='relative flex flex-col items-center'>
-      <img
-        src={src}
-        alt={alt || 'Image preview'}
-        className={cn(
-          'max-h-[50vh] w-auto object-contain dark:border dark:border-gray-700 dark:bg-gray-900',
-        )}
-      />
-    </div>
-  )
-}
-
-const VideoPlayer = ({ src, type }: { src: string; type?: string }) => {
-  return (
-    <div className='flex justify-center'>
-      <video
-        className='max-h-[50vh] max-w-full dark:border dark:border-gray-700'
-        controls
-        autoPlay={false}
-      >
-        <source src={src} type={type} />
-        <track kind='captions' src='' label='English' />
-        Your browser does not support the video tag.
-      </video>
-    </div>
-  )
-}
-
-const AudioPlayer = ({ src }: { src: string }) => {
-  return (
-    <div className='flex flex-col items-center justify-center rounded-lg bg-gray-100 p-6 dark:bg-gray-800'>
-      <MusicalNoteIcon className='mb-4 h-16 w-16 text-gray-400 dark:text-gray-500' />
-      <audio className='w-full' controls>
-        <source src={src} />
-        <track kind='captions' src='' label='English' />
-        Your browser does not support the audio element.
-      </audio>
-    </div>
-  )
-}
-
-const TextViewer = ({ content, extension }: { content: string; extension: string }) => {
-  const isCode = [
-    'js',
-    'jsx',
-    'ts',
-    'tsx',
-    'html',
-    'css',
-    'py',
-    'java',
-    'rb',
-    'go',
-    'rust',
-    'php',
-    'json',
-  ].includes(extension)
-
-  return (
-    <div className='relative overflow-hidden rounded-md'>
-      <div className='absolute right-2 top-2 z-10'>
-        {isCode ? (
-          <CodeBracketIcon className='h-5 w-5 text-gray-500 dark:text-gray-400' />
-        ) : (
-          <DocumentTextIcon className='h-5 w-5 text-gray-500 dark:text-gray-400' />
-        )}
-      </div>
-      <pre
-        className={cn(
-          'max-h-[50vh] overflow-auto p-4',
-          isCode
-            ? 'bg-gray-800 text-gray-100 dark:bg-gray-900'
-            : 'bg-gray-100 text-gray-800 dark:bg-gray-800 dark:text-gray-100',
-        )}
-      >
-        <code>{content}</code>
-      </pre>
-    </div>
-  )
-}
-
-const PDFViewer = ({ src }: { src: string }) => {
-  return (
-    <div className='flex flex-col items-center'>
-      <embed
-        src={src}
-        type='application/pdf'
-        className='h-[50vh] w-full dark:border dark:border-gray-700'
-      />
-      <a
-        href={src}
-        target='_blank'
-        rel='noopener noreferrer'
-        className='mt-2 flex items-center text-auto-drive-accent hover:underline dark:text-darkAccent'
-      >
-        <DocumentIcon className='mr-1 h-4 w-4' />
-        Open PDF in new tab
-      </a>
-    </div>
-  )
-}
-
-const PasswordModal = ({
-  isOpen,
-  onClose,
-  onConfirm,
-}: {
-  isOpen: boolean
-  onClose: () => void
-  onConfirm: (password: string) => void
-}) => {
-  const [password, setPassword] = useState('')
-
-  const handleConfirm = useCallback(() => {
-    onConfirm(password)
-    onClose()
-    setPassword('')
-  }, [onClose, password, onConfirm])
-
-  return (
-    <Dialog open={isOpen} onOpenChange={onClose}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle className='text-lg font-normal text-gray-900 dark:text-white'>
-            Enter Decryption Password
-          </DialogTitle>
-        </DialogHeader>
-        <div className='text-center'>
-          <input
-            id='password'
-            type='password'
-            placeholder='Enter Password'
-            className='w-full rounded border p-2 text-gray-900 dark:border-gray-700 dark:bg-gray-800 dark:text-white focus:outline-auto-explorer-primaryAccent'
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-          />
-          <button
-            className='mt-4 rounded-lg text-xs bg-auto-explorer-buttonLightFrom px-4 py-2.5 font-normal leading-4 text-white dark:bg-auto-explorer-buttonDarkFrom'
-            onClick={handleConfirm}
-          >
-            Confirm
-          </button>
-        </div>
-      </DialogContent>
-    </Dialog>
-  )
-}
+import { PasswordModal } from './PasswordModal'
+import { PDFViewer } from './PDFViewer'
+import { TextViewer } from './TextViewer'
+import { VideoPlayer } from './VideoPlayer'
 
 export type FilePreviewProps = {
   metadata: OffchainMetadata
   isAstral?: boolean
   isAutoDrive?: boolean
   network: NetworkId
-  loading?: boolean
+  loading: boolean
   file: Blob | null
   error: string | null
   isDecrypted?: boolean
@@ -186,7 +37,7 @@ export const FilePreview = ({
   isAstral = false,
   isAutoDrive = false,
   network,
-  loading,
+  loading = false,
   file,
   error,
   isDecrypted,
@@ -214,103 +65,60 @@ export const FilePreview = ({
     return data
   }, [file, metadata, gatewayUrl])
 
-  const preview = useMemo(() => {
-    const DirectGatewayLink = () =>
-      gatewayUrl && !metadata.uploadOptions?.encryption ? (
-        <div className='mt-2 flex justify-end text-sm'>
-          <a
-            href={gatewayUrl}
-            target='_blank'
-            rel='noopener noreferrer'
-            className={cn(
-              'flex items-center hover:underline',
-              isAutoDrive && 'text-auto-drive-accent dark:text-auto-drive-accent',
-              isAstral && 'text-auto-explorer-primaryAccent dark:text-auto-explorer-primaryAccent',
-            )}
-          >
-            <ArrowTopRightOnSquareIcon className='mr-1 h-4 w-4' />
-            View on gateway
-          </a>
-        </div>
-      ) : null
-
+  const renderErrorStates = () => {
     if (loading) {
-      return (
-        <div className='flex h-[50vh] items-center justify-center'>
-          <div
-            className={cn(
-              'h-12 w-12 animate-spin rounded-full border-b-2',
-              isAutoDrive && 'border-auto-drive-accent dark:border-auto-drive-accent',
-              isAstral &&
-                'border-auto-explorer-primaryAccent dark:border-auto-explorer-primaryAccent',
-            )}
-          ></div>
-        </div>
-      )
+      return <LoadingSpinner isAutoDrive={isAutoDrive} isAstral={isAstral} />
     }
 
     if (error) {
       return (
-        <div className='rounded-md border border-red-300 bg-red-50 p-4 text-red-700 dark:border-red-800 dark:bg-red-900/20 dark:text-red-400'>
-          Error: {error}
-          {gatewayUrl && <DirectGatewayLink />}
-        </div>
+        <ErrorDisplay
+          error={error}
+          gatewayUrl={gatewayUrl}
+          isAutoDrive={isAutoDrive}
+          isAstral={isAstral}
+        />
       )
     }
 
     if (decryptionError) {
       return (
-        <div className='rounded-md border border-red-300 bg-red-50 p-4 text-red-700 dark:border-red-800 dark:bg-red-900/20 dark:text-red-400'>
-          Decryption Error: {decryptionError}
-          <button
-            onClick={() => setModalOpen(true)}
-            className={cn(
-              'mt-2 rounded-lg px-4 py-2 text-sm font-semibold text-white',
-              isAutoDrive &&
-                'bg-auto-drive-accent hover:bg-auto-drive-accent/90 dark:bg-auto-drive-accent dark:hover:bg-auto-drive-accent/90',
-              isAstral &&
-                'bg-auto-explorer-primaryAccent hover:bg-auto-explorer-primaryAccent/90 dark:bg-auto-explorer-primaryAccent dark:hover:bg-auto-explorer-primaryAccent/90',
-            )}
-          >
-            Try Again
-          </button>
-          {gatewayUrl && <DirectGatewayLink />}
-        </div>
+        <DecryptionErrorDisplay
+          decryptionError={decryptionError}
+          gatewayUrl={gatewayUrl}
+          isAutoDrive={isAutoDrive}
+          isAstral={isAstral}
+          onRetry={() => setModalOpen(true)}
+        />
       )
     }
 
     // Handle encrypted files that haven't been decrypted yet
     if (metadata.uploadOptions?.encryption && !isDecrypted) {
       return (
-        <div className='flex flex-col items-center justify-center rounded-lg border border-gray-200 bg-gray-50 p-6 dark:border-gray-700 dark:bg-gray-800'>
-          <div className='flex items-center mb-4'>
-            <LockClosedIcon className='mr-2 h-6 w-6 text-auto-explorer-primaryAccent' />
-            <p className='text-sm font-normal text-gray-900 dark:text-white'>
-              No preview due to the file being encrypted.
-            </p>
-          </div>
-          <button
-            onClick={() => setModalOpen(true)}
-            className={cn(
-              'rounded-lg px-4 py-2.5 text-xs font-semibold text-white',
-              isAutoDrive &&
-                'bg-auto-drive-accent hover:bg-auto-drive-accent/90 dark:bg-auto-drive-accent dark:hover:bg-auto-drive-accent/90',
-              isAstral &&
-                'bg-auto-explorer-buttonLightFrom hover:bg-auto-explorer-buttonLightFrom/90 dark:bg-auto-explorer-buttonDarkFrom dark:hover:bg-auto-explorer-buttonDarkFrom/90',
-            )}
-          >
-            Decrypt File
-          </button>
-          {gatewayUrl && <DirectGatewayLink />}
-        </div>
+        <EncryptedFilePrompt
+          gatewayUrl={gatewayUrl}
+          isAutoDrive={isAutoDrive}
+          isAstral={isAstral}
+          onDecryptClick={() => setModalOpen(true)}
+        />
       )
     }
 
+    return null
+  }
+
+  const renderFileContent = () => {
     if (metadata.type === 'folder') {
       return (
         <>
           <FolderPreview metadata={metadata} network={network} />
-          <DirectGatewayLink />
+          <DirectGatewayLink
+            gatewayUrl={gatewayUrl}
+            isEncrypted={!!metadata.uploadOptions?.encryption}
+            isAutoDrive={isAutoDrive}
+            isAstral={isAstral}
+          />
         </>
       )
     }
@@ -318,7 +126,12 @@ export const FilePreview = ({
       return (
         <>
           <NoPreviewAvailable />
-          {gatewayUrl && <DirectGatewayLink />}
+          <DirectGatewayLink
+            gatewayUrl={gatewayUrl}
+            isEncrypted={!!metadata.uploadOptions?.encryption}
+            isAutoDrive={isAutoDrive}
+            isAstral={isAstral}
+          />
         </>
       )
     }
@@ -334,7 +147,12 @@ export const FilePreview = ({
       return (
         <>
           <ImageViewer src={fileData.uri} alt={fileData.fileName} />
-          <DirectGatewayLink />
+          <DirectGatewayLink
+            gatewayUrl={gatewayUrl}
+            isEncrypted={!!metadata.uploadOptions?.encryption}
+            isAutoDrive={isAutoDrive}
+            isAstral={isAstral}
+          />
         </>
       )
     }
@@ -342,7 +160,12 @@ export const FilePreview = ({
       return (
         <>
           <PDFViewer src={fileData.uri} />
-          <DirectGatewayLink />
+          <DirectGatewayLink
+            gatewayUrl={gatewayUrl}
+            isEncrypted={!!metadata.uploadOptions?.encryption}
+            isAutoDrive={isAutoDrive}
+            isAstral={isAstral}
+          />
         </>
       )
     }
@@ -353,7 +176,12 @@ export const FilePreview = ({
       return (
         <>
           <VideoPlayer src={fileData.uri} type={fileData.fileType} />
-          <DirectGatewayLink />
+          <DirectGatewayLink
+            gatewayUrl={gatewayUrl}
+            isEncrypted={!!metadata.uploadOptions?.encryption}
+            isAutoDrive={isAutoDrive}
+            isAstral={isAstral}
+          />
         </>
       )
     }
@@ -364,7 +192,12 @@ export const FilePreview = ({
       return (
         <>
           <AudioPlayer src={fileData.uri} />
-          <DirectGatewayLink />
+          <DirectGatewayLink
+            gatewayUrl={gatewayUrl}
+            isEncrypted={!!metadata.uploadOptions?.encryption}
+            isAutoDrive={isAutoDrive}
+            isAstral={isAstral}
+          />
         </>
       )
     }
@@ -393,7 +226,12 @@ export const FilePreview = ({
       return (
         <>
           <TextViewer content={textContent} extension={extension} />
-          <DirectGatewayLink />
+          <DirectGatewayLink
+            gatewayUrl={gatewayUrl}
+            isEncrypted={!!metadata.uploadOptions?.encryption}
+            isAutoDrive={isAutoDrive}
+            isAstral={isAstral}
+          />
         </>
       )
     }
@@ -401,7 +239,12 @@ export const FilePreview = ({
       return (
         <>
           <Arguments file={file} />
-          <DirectGatewayLink />
+          <DirectGatewayLink
+            gatewayUrl={gatewayUrl}
+            isEncrypted={!!metadata.uploadOptions?.encryption}
+            isAutoDrive={isAutoDrive}
+            isAstral={isAstral}
+          />
         </>
       )
     }
@@ -417,9 +260,21 @@ export const FilePreview = ({
         <div className='mt-4 text-sm text-gray-500 dark:text-gray-400'>
           {fileData.fileType || extension.toUpperCase()} file preview
         </div>
-        <DirectGatewayLink />
+        <DirectGatewayLink
+          gatewayUrl={gatewayUrl}
+          isEncrypted={!!metadata.uploadOptions?.encryption}
+          isAutoDrive={isAutoDrive}
+          isAstral={isAstral}
+        />
       </div>
     )
+  }
+
+  const preview = useMemo(() => {
+    const errorState = renderErrorStates()
+    if (errorState) return errorState
+
+    return renderFileContent()
   }, [
     loading,
     error,
