+++ packages/auto-design-system/src/components/common/FilePreview/types.ts
@@ -1,0 +1,6 @@
+// File data interface
+export interface FileData {
+  dataArrayBuffer: ArrayBuffer
+  fileName?: string
+  isEncrypted?: boolean
+}
